f
kuuga

+++ main.cpp
@@ -135,6 +135,6 @@ int main() {
 //    tinyJS.root->addChild("result", new Var(0));
     tinyJS.execute();
 
-    cout << tinyJS.root->findChild("result")->var->getInt() << endl;
+    cout << tinyJS.root->findChild("result")->var->getDouble() << endl;
     return 0;
 }
